@@ -2,7 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 
 class SellerModel{
   String email;
-  DocumentReference reference;
+  DocumentReference? reference;
 
   SellerModel({this.email = '', this.reference});
 
