@@ -9,7 +9,7 @@ import 'package:mobx/mobx.dart';
 class SplashPage extends StatefulWidget {
   final String title;
 
-  const SplashPage({Key key, this.title = "SplashPage"}) : super(key: key);
+  const SplashPage({Key? key, this.title = "SplashPage"}) : super(key: key);
 
   @override
   SplashPageState createState() => SplashPageState();
@@ -17,7 +17,7 @@ class SplashPage extends StatefulWidget {
 
 class SplashPageState extends State<SplashPage> {
 
-  ReactionDisposer disposer;
+  late ReactionDisposer disposer;
 
   @override
   void initState() {
