@@ -9,6 +9,14 @@ part of 'product_store.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
 
 mixin _$ProductStore on _ProductStoreBase, Store {
+  Computed<List<ProductModel?>>? _$computedProductsComputed;
+
+  @override
+  List<ProductModel?> get computedProducts => (_$computedProductsComputed ??=
+          Computed<List<ProductModel?>>(() => super.computedProducts,
+              name: '_ProductStoreBase.computedProducts'))
+      .value;
+
   final _$loadingAtom = Atom(name: '_ProductStoreBase.loading');
 
   @override
@@ -54,6 +62,21 @@ mixin _$ProductStore on _ProductStoreBase, Store {
     });
   }
 
+  final _$searchAtom = Atom(name: '_ProductStoreBase.search');
+
+  @override
+  String? get search {
+    _$searchAtom.reportRead();
+    return super.search;
+  }
+
+  @override
+  set search(String? value) {
+    _$searchAtom.reportWrite(value, super.search, () {
+      super.search = value;
+    });
+  }
+
   final _$getListAsyncAction = AsyncAction('_ProductStoreBase.getList');
 
   @override
@@ -61,12 +84,28 @@ mixin _$ProductStore on _ProductStoreBase, Store {
     return _$getListAsyncAction.run(() => super.getList());
   }
 
+  final _$_ProductStoreBaseActionController =
+      ActionController(name: '_ProductStoreBase');
+
+  @override
+  dynamic setSearch(dynamic value) {
+    final _$actionInfo = _$_ProductStoreBaseActionController.startAction(
+        name: '_ProductStoreBase.setSearch');
+    try {
+      return super.setSearch(value);
+    } finally {
+      _$_ProductStoreBaseActionController.endAction(_$actionInfo);
+    }
+  }
+
   @override
   String toString() {
     return '''
 loading: ${loading},
 products: ${products},
-lastProduct: ${lastProduct}
+lastProduct: ${lastProduct},
+search: ${search},
+computedProducts: ${computedProducts}
     ''';
   }
 }
