@@ -3,9 +3,9 @@ import 'package:e_commerce_app/app/shared/models/product_model.dart';
 
 abstract class IProductRepository{
 
-  Stream<List<ProductModel>> getAll(var productModel);
+  Stream<List<ProductModel?>> getAll(var productModel);
   Future save(ProductModel productModel);
   Future update(ProductModel productModel);
   Future delete(ProductModel productModel);
-  Future<QuerySnapshot> getAllPaginate(ProductModel productModel);
+  Future<QuerySnapshot> getAllPaginate(ProductModel? productModel);
 }
